requirements.txt
jax[cpu]>=0.4.0
flax>=0.7.0
optax>=0.1.0
numpy>=1.22.0
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
smbus2>=0.4.0